fix scene debug_cy + fix prototype rt_put_pixel + add comment and debug for cylinder + fix rotation value for now

+++ main_work/incs/rt_raytracer.h
@@ -6,7 +6,7 @@
 /*   By: ljohnson <ljohnson@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/03 12:20:32 by ljohnson          #+#    #+#             */
-/*   Updated: 2022/09/13 17:17:25 by ljohnson         ###   ########lyon.fr   */
+/*   Updated: 2022/09/13 17:21:25 by ljohnson         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -58,7 +58,7 @@ bool	rt_in_shadow(t_obj_data *obj_data, t_ray *l_ray, t_hit hit);
 
 /*-------------------- rt_ray_tracer.c --------------------*/
 
-void	rt_put_pixel(float x, float y, t_color color, t_mlx_data *mlx);
+void	rt_put_pixel(int x, int y, t_color color, t_mlx_data *mlx);
 t_ray	rt_create_ray(t_camera cam, float w, float h);
 bool	check_rotation_cam(t_ray *ray, t_camera cam);
 void	rt_ray_tracer(t_master *master);
